Aside: Make service a class
Actually, let's make the external dependency (SomeoneElsesLibrary.Service) a class rather than an object. This doesn't really have any bearing on the substance of the example, it just means that we won't have to digress into explaining how to reference the type of a singleton object before we get onto the problem we're actually trying to solve.