Step 2: Pass the service into the client constructor.
Here's a first improvement - we'll move the responsibility for instantiating the service object to the top level of an app, that way our client class is no longer responsible for instantiating its own collaborators. However, our class is still dependent on the external library at compile time, as it requires an object of a type Service, which is defined in the external library. Can we fix this?