Step 3: Removing the dependency on the Service class
So, in order to remove the dependency on the Service class from our Client, we can use the typeclass trick ourselves. We define a Sender typeclass, which describes the interface we're interested in from the external service, and provide a typeclass instance which wraps is (Adapters.SomeoneElsesServiceSender). As long as this adapter is in scope at the call site of the Client constructor, the Client class need only depend on our (abstract) Sender trait, and now no longer needs to reference the third-party Service at all! This is good news as we can now test it with our own mocked out Sender, rather than relying on the concrete third-party implementation.

We could, in theory, plug in any other Sender implementation too, but there's a bit of a hiccup here. Our Client class, and our Sender trait, both still rely on the third-party library's Response type-class, so our client still has a dependency on the third-party library, and further, any other Service implemetations would do too! This isn't ideal (and to some extent defeats the point of factoring out the Service dependency in the first place). Can we do better?