(Failed) Step 4a: Just remove the references to the typeclass
We might think, since neither the client's send method, or the Sender's SendWith method reference the TheirResponse trait, then we can just remove the constraint from their type signatures and be done with it!
This however, is sadly not the case. To see why, look at our definition of sendWith in the SomeoneElsesServiceAdaptor. The method body can't fulfil the contract given by its own type signature, as service.send(a) is only defined for A with an instance of TheirResponse, which is a subset of our declared type A! This won't compile, and we'll have to try something else...