(Failed) Step 4b: Using the service adapter to hide the third party typeclass.
We could try and reintroduce the type constraint on sendWith inside the Adaptor - since it already depends on the concrete Service, it'd make sense to try and isolate all the dependencies on the third-party library here, right?

We're getting there, but it's not quite that simple. The type signature of the adaptor's sendWith method no longer matches the (generic) signature of our Sender typeclass, so it's no longer an instance of that type. This still won't compile, as we've just pushed the problem up a level! Clearly, we need a way to constraint the type of a sendable thing in a way that doesn't depend on the third-party library.