(Failed) Step 4c - Define our own Sendable class.
Thinking about how this type error is just percolating up through the system we get to the root of the problem. At the moment, the send method on our client takes *any* type - we haven't defined this interface at all ourselves, so the restriction imposed by the other library is actually the narrowest in our whole system! If we were a bit more selective about our inputs, we wouldn't have this problem. In a sense, we've had the whole problem backwards. Rather than deciding on our own inputs and outputs, we've just exposed whatever the library we rely on can deal with. Let's remedy this, by explicitly stating that our client can send Strings and Ints, using another typeclass, which we'll call 'sendable'.

We're still not quite there though - our Sendable typeclass (which is just a union of Strings and Ints) is now a subset (although entirely coincidentally, not a proper subset - there's no need for the members of the two classe to be exactly equal) of the library's TheirResponse type. However, the compiler doesn't know this, so we're still stuck. Can we enforce this constraint within our adaptor?